pyproject.toml
[tool.poetry]
name = "project24-1"
version = "0.1.0"
description = ""
authors = ["TolyaLaz <toljik1@mail.ru>"]
readme = "README.md"
packages = [{include = "project24"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.2"
psycopg2 = "^2.9.9"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"
serializers = "^0.2.4"
python-dotenv = "^1.0.1"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
ipython = "^8.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
